% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f1.R
\name{payments_by_drg_code2}
\alias{payments_by_drg_code2}
\title{Boxplot of Payments by DRG Code}
\usage{
payments_by_drg_code2(data, type)
}
\arguments{
\item{data}{a datafame}

\item{type}{a string indicating which type of payment to plot}
}
\value{
returns a boxplot of payments by DRG code
}
\description{
Boxplot of Payments by DRG Code
}
\examples{
payments_by_drg_code2(drg, "Average Medicare Payments")

}
