% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_payments.R
\name{stats_payments}
\alias{stats_payments}
\title{Statistics for Average Medicare Payments}
\usage{
stats_payments(data, stat)
}
\arguments{
\item{data}{a dataframe. Only the drg dataframe is applicable.}

\item{stat}{a string specifying which statistic to return. Must be one of "mean", "median", "sd", or "all".}
}
\value{
A tibble with the calculated statistic(s) for average Medicare payments, grouped by DRG code.
}
\description{
Calculates the mean, median, standard deviation, or all of these statistics for average Medicare payments
based on \code{stat} over all DRG codes.
}
\examples{
stats_payments(drg, "mean")

stats_payments(drg, "all")

}
