% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/payments_by_drg_code.R
\name{payments_by_drg_code}
\alias{payments_by_drg_code}
\title{Boxplot of Payments by DRG Code}
\usage{
payments_by_drg_code(data, type)
}
\arguments{
\item{data}{a dataframe}

\item{type}{a string indicating which type of payment to plot}
}
\value{
Returns a boxplot of payments by DRG code
}
\description{
Boxplot of Payments by DRG Code
}
\examples{
payments_by_drg_code(drg, "Average Medicare Payments")

}
