% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/payments_by_drg_code.R
\name{payments_by_drg_code}
\alias{payments_by_drg_code}
\title{Boxplot of Payments by DRG Code}
\usage{
payments_by_drg_code(data, type)
}
\arguments{
\item{data}{a dataframe. Only the drg dataframe is applicable.}

\item{type}{a string specifying which type of payment to plot. Must be one of "Average Covered Charges", "Average Total Payments", or "Average Medicare Payments".}
}
\value{
A boxplot of specified type of payments by DRG code.
}
\description{
Creates a ggplot boxplot of payments for each DRG code in the drg dataframe.
}
\examples{
payments_by_drg_code(drg, "Average Medicare Payments")

payments_by_drg_code(drg, "Average Covered Charges")
}
