% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_payments.R
\name{stats_payments}
\alias{stats_payments}
\title{Statistics for Average Medicare Payments}
\usage{
stats_payments(data, stat)
}
\arguments{
\item{data}{a dataframe}

\item{stat}{a string name for which statistic to return}
}
\value{
A statistic for average medicare payments grouped by DRG code
}
\description{
This function returns the mean, median, or standard deviation for average medicare payments based on the \code{stat} grouped by DRG code.
}
\examples{
stats_payments(drg, "mean")


}
